Fect(Contactupdate): Added REST endpoint for making contact favourite
Updated Contact model
Added PUT REST endpoint
Update individual contact added to Business logic
Methods added to Contact detail Contract to handle updated contact

+++ app/src/main/java/com/example/aditya/gojek/data/model/Contact.java
@@ -18,7 +18,7 @@ public class Contact implements Comparable<Contact> {
 
     @StorIOSQLiteColumn(name = DatabaseContract.Contacts.COLUMN_CONTACT_ID, key = true)
     @StorIOContentResolverColumn(name = DatabaseContract.Contacts.COLUMN_CONTACT_ID, key = true)
-    @SerializedName("id") Long id;
+    @SerializedName("id") int id;
 
     @StorIOSQLiteColumn(name = DatabaseContract.Contacts.COLUMN_FIRST_NAME)
     @StorIOContentResolverColumn(name = DatabaseContract.Contacts.COLUMN_FIRST_NAME)
@@ -60,11 +60,11 @@ public class Contact implements Comparable<Contact> {
         this.firstAlpha = firstAlpha;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
